refactor: separation of forms and result folders.
