+++ src/components/Forms/Auth/ChangePassword/ChangePassword.props.ts
@@ -1,4 +1,4 @@
-import { IAuthResultProps } from "@components/AuthResult/AuthResult.props";
+import { IAuthResultProps } from "@components/Results/AuthResult/AuthResult.props";
 
 
 export interface IChangePassword extends Omit<IAuthResultProps, 'buttonTitle'> {
