+++ src/components/Forms/Auth/ConfirmEmail/ConfirmEmail.props.ts
@@ -1,4 +1,4 @@
-import { IAuthResultProps } from "@components/AuthResult/AuthResult.props";
+import { IAuthResultProps } from "@components/Results/AuthResult/AuthResult.props";
 
 
 export interface IConfirmEmailProps extends Omit<IAuthResultProps, 'buttonTitle'> {
