+++ src/components/Forms/Auth/RegistrationForm/RegistrationForm.tsx
@@ -2,7 +2,7 @@ import { GooglePlusOutlined } from "@ant-design/icons";
 import { Button, Form, Input, Tabs, Image, Grid } from "antd";
 import { Link, Navigate, useLocation } from "react-router-dom";
 import FullLogoIcon from "/icons/full-logo-icon.svg";
-import { Paths } from "../../../routes";
+import { Paths } from "../../../../routes";
 import { useRegistrationMutation } from "@services/auth-service";
 import { useEffect, useState } from "react";
 import { IAuthForm } from "@interfaces/auth.interface";
