+++ src/components/Results/AuthResult/AuthResult.props.ts
