+++ src/components/Results/AuthResult/AuthResult.tsx
